clean code
organised code into different files for better handling
